Patch potential infinite loop
An infinite loop could be caused if, for some reason, `Featurette.load`
is called as a result of a featurette's constructor. This is because the
featurette's constructor is, necessarily, called before it can be added
to Featurette.featurettes. Adding a placeholder before the class is
instantiated is ugly, but effective.